Rename: 슬라이드 추가를 위한 preview디렉토리 생성 및 setPreview.js이동
